devlabs-rvision-hack-ner-service/rvision-hack-main/requirements.txt
pika==1.1.0
psycopg2-binary==2.8.6
requests==2.22.0